telegram_bot/pyproject.toml
[tool.poetry]
name = "rent-service-bot"
version = "0.1.0"
description = "Telegram bot for rent monitoring"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.3.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.9.1"
isort = "^5.12.0"
flake8-pyproject = "^1.2.3"

# formatting
[tool.poe.tasks.format]
shell = "black src && isort src && flake8 src --format='telegram_bot/%(path)s:%(row)d:%(col)d: %(code)s %(text)s'"

[tool.flake8]
max-line-length = 120
max-complexity = 15
ignore = "E501, W503, E722, E203"
per-file-ignores = "__init__.py:F401"

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2

[tool.pylint.format]
max-line-length = "120"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 